Updated openstack/openstack Project: openstack/python-cinderclient  729b8f8474e6653cdd9eb34c6c2b558024c1ddcb
Don't reset volume status when resetting migration status

In case of failed volume migration, status of the volume is
still in-use and the migration status is set to error.

Current reset-migration-status command resets not only
migration status but also volume status. However the volume
status should not reset because the volume is still attached.

Closes-Bug #1552058
Change-Id: I9a8a5ed6a00bdcffecbf98862fe60aee373f5e9b