Updated openstack/openstack Project: openstack/neutron-vpnaas  0171e1a177cc3d0481cd40e79d41f3e0afa9c0ff
Put py34 first in the env order of tox

To solve the problem of "db type could not be determined" on py34 we
have to run first the py34 env, then, run py27. This patch puts py34
first on the tox.ini list of envs to avoid this problem to happen.

Change-Id: I9457f7e05adb77d4b034ef01e7e51d95fc066013
Closes-bug: #1489059