Updated openstack/openstack Project: openstack/cinder  820e06a674791326c9c586d6ec6f79c85291e38d
Add backend id to Pure Volume Driver trace logs

With multi-backend deployments it was very hard to follow which backend
was making which calls. When there was an error it was often unhelpful
to look at the tracing to know which backend had the problem.

This will now print out the active backend id with the tracing log
messages.

Change-Id: Iaa1ec1c9027f6ba8aeb2574b92a416ae4bd86455
Closes-Bug: #1552966