Updated openstack/openstack Project: openstack/horizon  198e07413f5535f3209ade70d4100b273fd8ed77
Hamburger navigation now sits above containers

In the material theme, if the screen is too narrow, the navigation pane
becomes a collapsable "hamburger-menu" style button.  Currently, on the
containers page, when that button is pressed the navigation pane is
below the containers list.  This makes navigation impossible.

See screenshot here: http://i.imgur.com/uzdX66E.png

Adjusting the z-index of the container down from 10 to 2 aleviates this
issue.  Obviously the containers page is being reworked elsewhere; so
keeping this change as small as possible is appropriate.

Change-Id: I220d2f8a87e9a55b95884251b3368abe1167cafa
Closes-Bug: 1552898
Partially-implements: blueprint horizon-theme-css-reorg