Updated openstack/openstack Project: openstack/nova  832bb880fd6e74ca81b3ae8ed1a9c59366010807
Check 'destination_type' instead of 'source_type' in _check_and_transform_bdm

In compute.api._check_and_transform_bdm() we have a logic to
avoid boot instances with both image-ref and a volume named
as 'vda' is supplied. Currently, we check the bdm's 'source_type',
but infact we should check its' 'destination_type' as this
shows it is a cinder volume.

Change-Id: I1fe2cf7c6655e0e0c61371c6d7379ecfc7071cec
Closes-Bug: #1522329