Updated openstack/openstack Project: openstack/nova  f566aad1757cbe14b7ad43a45598a634cd0bf6b1
Add specific method to lazy-load instance.pci_devices

The pci_devices field is generically lazy-loaded today by
getting the instance record from the database with a join
on the pci_devices table. We can make this smarter by just
getting the PciDeviceList via the instance uuid directly
so we don't need to do the join and query on the instances
table.

Change-Id: I90f6c7a17bb773798ae77d19d744dcac02de215c
Related-Bug: #1540526