Updated openstack/openstack Project: openstack/keystone  bc32abc526983b9060f700c846c9fd6e6597ca56
Split resource backend tests

Where to put a new test ? Is this behavior covered yet ? These and
other questions require understanding the test coverage to be answered.

test_backend.py is a huge test file, making it hard to understand what
cases are currently covered for each backend.

This chain of changes proposes to split it by the backends it covers;
the tests will be placed into 'unit/{backend}/test_backends.py'.

This change splits the tests for the resource backend.

Change-Id: Ibf09008e558b2cc9e53301c692737d305fb31695