Updated openstack/openstack Project: openstack/keystone  85b770e5bac1c296dea572a92a49740b42389be1
Split token backend tests

Where to put a new test ? Is this behavior covered yet ? These and
other questions require understanding the test coverage to be answered.

test_backend.py is a huge test file, making it hard to understand what
cases are currently covered for each backend.

This chain of changes proposes to split it by the backends it covers;
the tests will be placed into 'unit/{backend}/test_backends.py'.

This change splits the tests for the token backend.

Change-Id: If14b8e7ec0477ca6126e65acf4d7108d6ce7ed19