Updated openstack/openstack Project: openstack/keystone  cc3e391407a8dfada5f25a66e9b5960e52e9ddf8
Split trust backend tests

Where to put a new test ? Is this behavior covered yet ? These and
other questions require understanding the test coverage to be answered.

test_backend.py is a huge test file, making it hard to understand what
cases are currently covered for each backend.

This chain of changes proposes to split it by the backends it covers;
the tests will be placed into 'unit/{backend}/test_backends.py'.

This change splits the tests for the trust backend.

Change-Id: I88b9d3976ab1e35e32a493f07c8f5be1ece7f531