Updated openstack/openstack Project: openstack/cinder  e7f7cccf5a5f3f86b916d43e7bc4f3b828f6c878
CONF add suppress_requests_ssl_warnings

Many cinder drivers now seem to use the python requests library
to talk to their arrays.  One of the joys of the requests library
is that it correctly complains when it connects to a service where
the SSL certificate is bogus or out of date.  While this is a much
needed warning to raise in production environments, it's often a
useless warning for developers and or for POC deployments of Cinder.

This patch adds a CONF setting that allows the volume manager to
disable all requests ssl warnings for each driver instance.
This greatly cleans up the c-vol logs and makes them readable.

By default the CONF setting is False, or disabled.  So the admin
has to manually edit the cinder.conf file and turn this on.

DocImpact
Change-Id: I2f3d21d828aea9a04cf006dcd471ece668bfd947