Updated openstack/openstack Project: openstack/keystone  4aeb13d13ddbb69c78b9b3ed31ec0a20519bb236
Split catalog backend tests

Where to put a new test ? Is this behavior covered yet ? These and
other questions require understanding the test coverage to be answered.

test_backend.py is a huge test file, making it hard to understand what
cases are currently covered for each backend.

This chain of changes proposes to split it by the backends it covers;
the tests will be placed into 'unit/{backend}/test_backends.py'.

This change splits the tests for the catalog backend.

Change-Id: I7a512e7662deac911c40763bd5f4c139b093a9e2