Updated openstack/openstack Project: openstack/api-site  6b33a42be685fca71b81fe17882bd1f1c61b93d2
[Orchestration API v1] Clarify language about stack names

As was pointed out on the original review[1] that introduced it, the
language about unique stack names becoming available again only after an
"unspecified amount of time" is not based on any actual restriction that
exists in Heat. Stacks are soft-deleted and therefore available to query
(by UUID) even after they have been deleted until such time as they are
purged from the database (an interval which is controlled by the operator,
and hence could be considered unspecified), but the name becomes available
as soon as the delete has successfully completed.

[1] https://review.openstack.org/#/c/131277/1/api-ref/src/wadls/orchestration-api/src/v1/orchestration-api.wadl

Change-Id: Ie002e2537bcff53fcdbc60d6ac59fb0dd2edacd7