Updated openstack/openstack Project: openstack/keystone  4287b79dac14f1120f845f21c78c2e0b1b6f9f67
Split identity backend tests

Where to put a new test ? Is this behavior covered yet ? These and
other questions require understanding the test coverage to be answered.

test_backend.py is a huge test file, making it hard to understand what
cases are currently covered for each backend.

This chain of changes proposes to split it by the backends it covers;
the tests will be placed into 'unit/{backend}/test_backends.py'.

This change splits the tests for the identity backend.

Change-Id: I5d765adaf592bbcb2f8414bb636f815ae06bd540