Updated openstack/openstack Project: openstack/keystone  7f22dad8f953822d368948e287c97db4477b07b7
Remove _disable_project from the resource API

The _disable_project method in was only there to act as a method we could put a
decorator on for notifications. The notification decorator has been replaced
with an inline statement for sending the notification. This means we no longer
need to have a specific method for the purpose of a decorator, so we can remove
_disable_project.

Change-Id: I352e7e1edb2fad71ff940e0db8b6f1dd1697df77