Updated openstack/openstack Project: openstack/python-openstackclient  0b2c4b1f327dddce0306637e7128bad6eb43fc54
Refactor security group set to use SDK

Refactored the 'os security group set' command to use the SDK
when neutron is enabled, but continue to use the nova client
when nova network is enabled.

This patch set also fixes a compute bug which ignores name
and description when set to an empty value.

Change-Id: I4225179dca4aedf799e1656ec49236bdedc5e9bd
Partial-Bug: #1519511
Implements: blueprint neutron-client