Updated openstack/openstack Project: openstack/nova  5b972c62bae1463a48842c26af3b686846972f1d
VMware: Use actual VM state instead of using the instance vm_state

In review https://review.openstack.org/169732 it was suggested that
we use the actual VM state instead of relying on what the instance
vm_state is.

So prior to attaching or detaching an IDE disk we check the VM state.
This is due to the fact that it does not support hot plugging.

Change-Id: Ibc104a20c528ac6c8f84351f518f3dd7a185794f