Updated openstack/openstack Project: openstack/neutron  284f507adaa2af7691e49327cdfc2baba9268455
Delay description association proxy construction

Association proxies cannot be constructed in the mixin because they
will be shared between all models that inherit from it. This means
they all share one and it will effectively be broken for all but the
first to load it.

This changes it to a declared attr like the others to make sure it's
not constructed until each model inherits from it.

Change-Id: I3e1aa21cd02c0f116cc6dee5e0577988ec37f767
Closes-Bug: #1553689