Updated openstack/openstack Project: openstack/keystone  2a24e359c6af26e4da6c2b4b2dc40a958acc59ad
Correct create_project driver versioning

create_project works differently in the V9 resource driver than it
did in V8. The V9 driver allows passing a null domain_id whereas the
V8 driver would fail.

Since the methods are different, create_project should not be
defined in ResourceDriverBase but in ResourceDriverV8 and
ResourceDriverV9. Also, its behavior should be documented and tested.

Also, the existing documentation was incorrect. It said that Conflict
would be raised if there was already a project with the same name but
it's only if the same name in the same domain.

Change-Id: Ib276310056f8826b97ef2b5f94b91840c46bb761