Updated openstack/openstack Project: openstack/nova  3781ef8e0cda3781b9e876279d02ce3fd35a4f9b
Ironic: Clean up if configdrive build fails

The configdrive builder does a number of things; for instance, shelling
out to mkisofs or reaching up to the top cell for SSH keys. These have
ways of failing. If they do fail, we need to cleanup what we've done in
the deploy so far (instance fields on the node in ironic, firewalls,
vifs, etc).

Wrap the configdrive build call in a try/except to catch any errors,
clean up, and re-raise the error.

Change-Id: I81ec4c54437d92c4c091646c0b3f5d948b82b52b
Closes-Bug: #1552466