Updated openstack/openstack Project: openstack/cinder  e26155183e8becc86b91d875a316590d3954c0ce
Readd iscsi_target table

Rolling upgrades was broken when iscsi_target table was dropped on
https://review.openstack.org/268320

We cannot stop using a table and drop it in the same release for rolling
upgrades to work, we have to stop using it in one release and then drop
it in the next or in the post rolling upgrade mechanism (which is still
not in place).

So this patch fixes this by removing the dropping and adding another
migration that ensure that the table is really there.  That way we can
be sure that anyone using M will have the table, which then will get
dropped in N.

Closes-Bug: #1553079
Change-Id: I26586549485a8d745a25161d97d33426fdd52576