Updated openstack/openstack Project: openstack-infra/project-config  0d09da872a329cefe998131696a35689644356bd
Set tags-only docs publishing for client jobs, not server jobs

Recently https://review.openstack.org/279238 merged, resulting in
docs for server/service repositories being published in a way that
suits the library/client repo publishing process (ie only the
latest tag is published).

This approach was meant to be implemented for the *client* jobs,
not the *server* jobs.

This patch corrects the behaviour to match the intent.

Change-Id: Iec0bb7ae930560d6617b58859c38cde0ab11eeb9