Updated openstack/openstack Project: openstack/openstack-manuals  30c8045b78de6080198b19d3f629416ecf45cbe6
Do not translate Config Reference

The Config Reference has no translations at all currently and contains a
lot of generated content. Do not add it to our translation list for now,
we can revisit it again if there is real demand and time to translate
it. Right now it just creates needless imports.

Change-Id: I94d52c1c03d5909e4c88e8aea898febe2ff222b6