Updated openstack/openstack Project: openstack/nova  240b3507e218c3f8d59e12dd067790eafec1e969
Provide correct connector for evacuate terminate

During evacuation a local connector is built. This is the
wrong connector to use for cinder terminate_connection.
In order to fix this, store the initial connector with
the BDM connection_info. Use the stored connector when
we detect that we have this wrong host situation.

This fix does not work for existing attachments
(made prior to this patch) because existing attachments
don't have the connector stashed in the bdm.connection_info.
In cases where the original connector was not saved, leave
the behavior as-is.

Change-Id: I793f2996fc0af1c321a240ad9348dc9bce816030
Partial-Bug: #1522496