Updated openstack/openstack Project: openstack/cinder  de88c809bf4b56697c1cbdbfb0d76d4320189309
Make nullable of fields in db model and object match

Currently there are some fields whose nullable in db
model and object don't match. This may lead object query
fails.

For example, for volume, if attach_status is nullable in
db schema, but is not nullable in object schema.
When writting data to db using db interface, the attach_status
can be null. But later when reading data from db to construct
volume object, it raises exception. As a result, all the
query related to the volume fails.

This patch is to make all the fields in corresponding db
model and object match.

Change-Id: I84fb16f0a67f9a1f7a09e91bb8ebe6298f3f49c4
Closes-Bug: #1549669
Related-Bug: #1549673