Updated openstack/openstack Project: openstack/tripleo-heat-templates  98224f0b511f5387a64e1871f2142236ee1c7712
Add IPv6 versions of the Controller NIC configs

This change adds Controller NIC configs for the sample NIC config
templates that are compatible with IPv6 on the External network.
These controller-v6.yaml templates include a default route for IPv6
on the External network, and a default route for IPv4 on the Control
Plane. The Heat parameters ExternalNetworkDefaultRoute and
ControlPlaneDefaultRoute are used to set these values.

Change-Id: Ifed8cb359eae1d9d623d3eb2fe40ea8a0d1d889a