Updated openstack/openstack Project: openstack-infra/project-config  547aa9c06bfa8414cc9ffeb2de25761c3132864e
Fix cache-bindep breakage on Fedora / CentOS

Currently, we don't have 100% gate coverage for
other-requirements.txt. As such, it is possible for our DIB builds to
fail on bindep.  So, we only cache packages for Ubuntu Trusty ATM
until we can enable 100% coverage for centos and fedora for
other-requirements.txt per project.

Change-Id: I0f06619224ab59bc6165316226abeccb75c101cb
Signed-off-by: Paul Belanger <pabelanger@redhat.com>