Updated openstack/openstack Project: openstack-infra/project-config  d0ac1478e67cf406090221c3a38099b64e9b61c8
Remove non-voting jobs from the gate pipeline.

Non-voting jobs should not be in the gate pipeline. Currently
there are a number that are non-voting in the gate pipeline.

This patch removes those jobs from the gate pipeline. Most jobs
are already running in the check pipeline where they are fine as
non-voting jobs. In one instance, on the devstack repo, the
grenade-dsvm-forward job was only in the gate pipeline so it was
moved to the check pipeline instead of the gate pipeline.

Change-Id: Ia14d724a7052c35a7bab31c2975876363848d98b