Updated openstack/openstack Project: openstack/tripleo-heat-templates  89b1177f58f1e9a40894b731734dbd2b67f5aefe
Fix rabbit_hosts list for glance-api for IPv6

Previously we were always appending the :port suffix to the list
of rabbitmq nodes but the syntax was invalid for IPv6.

This change wires rabbit_hosts from the templates as it happens
already for the other services. Port can be customized using
rabbit_port.

Change-Id: Iecc7a97d46d7de17e85398c57996c104c9125b0e