Updated openstack/openstack Project: openstack/python-novaclient  58356df3d91aa4ec1731bbbbaa5f1aa18ecb95a6
Remove console expectation from NMI tests

The original tests expect the server console contains a message
  "Uhhuh. NMI received for unknown reason"
but the expectation is wrong because this message is fallback one
when disabling crash dump, that means the test expects the crash
dump is disable. That is not the purpose of this NMI feature.

In addition, this message is output on x86 cpu architecture only
because the message is implemented in
  https://github.com/torvalds/linux/blob/master/arch/x86/kernel/nmi.c#L304
So if we use the other cpu architectures on the test, this test will
fail.

Let's remove the expectation from the tests.

Change-Id: I62fe2e00433fd216def25d99a2f20d17b988fb84
Closes-Bug: #1553597