Updated openstack/openstack Project: openstack-infra/storyboard-webclient  13bf67d958893cdab13da4e6c8f69855d2bd3bdb
Add a quote button to comments

This commit adds a quote button to comments, which appends a quoted
version of the comment to the content of the "new comment" text box.

This makes it much easier to preserve markdown and allows quoting in
one click instead of having to copy and paste.

Change-Id: I0f9b8ee0dbfeca57cc4c65d8b7a55f95e027dce4