Updated openstack/openstack Project: openstack/tempest  df1cb4f41163e4ddab2ba50215c9839afa3e05c6
Making Keystone clients consistent

This patch is intended to make consistency between Keystone V2 and V3
clients. If a client exists in V2 and V3, e.g roles client, users client,
suffix v3 is used for the corresponding V3 clients. If we deal with a client
that exists only in V3, e.g regions client, no suffix is used because we
don't have the same client in V2.

Change-Id: I976db1540035bf36c9c50d9d10987dd01b072e43