Updated openstack/openstack Project: openstack/security-doc  6567f1b9850f662e0bb9bf39f7c2ba64d66c210f
Add note about different cloud user types

As per bug, add a note about the cloud user types used in the doc.
Instead of describing them, a link to the glossary is all that is
required. Further work will need to be done to ensure the correct usage
is used throughout the book.

Change-Id: I47234ab4ce6ea2b26373ea38f79401be620a08ec
Partial-Bug: #1342368