Updated openstack/openstack Project: openstack/neutron  27c6cf94f2bb93a590fb670572d188ecfea5a9fa
Use floating IP to connect different address scopes

Currently, floating ip could be used to connect internal subnet and external
subnet in different address scopes. But the internal subnets that are in the
same address scope of external network are isolated. This is diagramed as
follow:
+---------------------+       +------------------------+
|   address scope 1   |       |    address scope 2     |
|                     |       |                        |
|  +----------------+ |       |   +---+----------+-+   |
|  |internal network| |       |   |external network|   |
|  +------------^---+ |       |   +-------^--------+   |
|               |     |       |           |            |
|       +-------+--+  |       |   +-------+----+       |
|       |fixed ip A+-------------^+floatingip B|       |
|       +----------+  |       |   +------------+       |
+---------------------+       |                        |
                              | +--------+  +--------+ |
                              | |internal|  |internal| |
                              | +-----+--+  +----+---+ |
                              +------------------------+

This patch will make the fixed ip A be able to connect internal networks in
address scope2. And internal networks in address scope2 are able to connect
fixed ip A by visiting floating ip B.

Change-Id: If3b0b49a994e8c604809e524dca6f4afe5b626f3
Closes-Bug: #1543885