Updated openstack/openstack Project: openstack/zaqar  3308af985be271245a2b739a1134bb1317e4c7c0
Renew subscription on update with TTL

Now when we update subscription and specify TTL, subscription is not
renewed. It's life is not extended, and it's expiration date is still
old.

Because of that, it's impossible to extend life of subscription without
recreating it. But during subscription recreation some messages may not
be sent to subscriber, because there will be a moment when subscription
does not exist, which is bad.

This patch makes subscriptions renewable by PATCH method, when TTL is
specified in the body. In such case PATCH method sets subscription TTL
to a new value and sets expiration date to a recalculated value: current
time + TTL.

This patch also adds two functional tests for testing TTL-related things
in subscriptions. This increases functional testing time by ~2-4
minutes.

APIimpact
Closes-Bug: #1552866
Co-Authored-By: wangxiyuan <wangxiyuan@huawei.com>
Co-Authored-By: Eva Balycheva <ubershy@gmail.com>
Change-Id: I048c37d6485ff27cf18ed27988874e470301fb57