Updated openstack/openstack Project: openstack-infra/project-config  724dad6b5f1a35c717f7190f1f21d00eeda2c8f0
Remove revoke-sudo from storlets, trove functional

The I670641b676231233f1dd583ea36421f8ffef8f8e change incorrectly
added a revoke-sudo macro to the gate-storlets-functional and
gate-trove-buildimage-{datastore} jobs which did not previously have
it, so remove it to get them working correctly again. It also added
one to dev-zuul-coverage, but I expect that was missing as an
oversight so not undoing it here.

Change-Id: Ic427c053602017da3d77b01193b4215f57de6fdd