Updated openstack/openstack Project: openstack-infra/shade  543954dd056527a4ff5164d0aec3e791c89b0550
In the service lock, reset the service, not the lock

It turns out that when you're in a context manager that has gotten the
lock, attempting to set that lock to None will fail. It's also not the
thing we actually want to do. What we WANT to do is reset the _service_
so that we'll ensure a new service object gets created.

Change-Id: Ic39ef0418dcd5c0cbbe710ac1c31e58b744f56b5