Updated openstack/openstack Project: openstack/heat  0e16abc96ef16bb11093b4d29df9363c61f04014
Make auth_url lookup dynamic

If _get_auth_url() is run on heat-api startup, it can cause heat-api to crash
in situations where Keystone is not running, yet or temporarily unavailable.
This patch converts the auth_url attribute into a property method that is only
run when it is needed, thus preventing this race condition.

Change-Id: Ife6d9e51ea9647e4658105c016867efe769e5363
Closes-Bug: #1550284