Updated openstack/openstack Project: openstack/tempest  997b2605be8557a028a3e93b03dee1202ea5d2af
Add OS-INHERIT of keystone v3 in api tempest

OS-INHERIT of keystone v3 does not exist api client and test.
It is an important function to inherit role assignments from
either their owning domain or projects that are higher in the hierarchy.
I have added OS-INHERIT of keystone v3 in api tempest.

Change-Id: Ic67ce4da124ea759fdaaf52aef17e8a1581bc3a8
Closes-bug: #1517722