Updated openstack/openstack Project: openstack/heat  295327453f3ea74262e91b0a1c09661369e38a20
Fix potential race condition in lbaas v2 logic

In the event that multiple resources under the same loadbalancer get
the affirmative that the loadbalancer is no longer updating before
either of those resources actually perform their action, neutron will
throw a PENDING_UPDATE error when the second resource attempts to
perform its action.

This patch implements the "better to ask forgiveness than to ask for
permmission" adage. Instead of checking whether the loadbalancer is
ready for an action to be performed, we should just try the action.
If it fails, we catch the error and try again later.

Closes-Bug: #1552724

Change-Id: I9f6e164ed838aa298a41b4ed2a6061a927787a4c