Updated openstack/openstack Project: openstack-infra/system-config  0b6a80ad1d3b9b6d13a2ea5179648e2cd165186d
Add vexxhost mirror to documentation and cacti

Add the new mirror.ca-ymq-1.vexxhost.openstack.org host to our
bandersnatch documentation and also correct a typo in another entry
there. Further, add it to the cacti monitoring list and remove the
entry for the dead HPCloud mirror while we're in the neighborhood.

Change-Id: I4be0dfc0099336773ea94bd940e55c4b0f7d036f