Updated openstack/openstack Project: openstack-infra/devstack-gate  ceb9cd120b3f2129d4322c0fc283971b4dc6aec8
Run tests in venv

Previously we had wrongly assumed that python yaml package
will be available in the images and the run-tests.sh script
ran a set of things which included a script that used
the pyyaml package. With the recent bindep changes, this
assumption broke and the gate-devstack-gate-unittests
failed. So we need to install the pyyaml package and then run
the tests in this venv for run-tests.sh to work.

Change-Id: I541c8add8041348a259b3b789d24cf24ba122ce0