Updated openstack/openstack Project: openstack/python-troveclient  3c71e52b408bbd6c33fca85188db2c8f156fb341
Revert "Time to get rid of any last vestiges of slave_of"

This reverts commit a5234f0a07230e0d8195687e033adfe4cc0ad522

The change broke trove on stable/liberty since it's a backward
incompatible change.

To avoid capping python-troveclient<2.1.0 in
stable/liberty global-requirements, we have to revert this
change, blacklist the 2.1.0 version in g-r master and
stable/liberty, and then release 2.2.0.

And the slave_of code can't be removed until the last
supported stable branch that uses it is end of life, which
is liberty-eol. For now we keep the slave_of argument in
the instance.create method, but emit a deprecation warning
if it's specified, and we never pass it to the trove API
for the create or edit operations since the API removed
slave_of in 0c94a5bae0aacff7fde9a9742b7d526096e51767.

Change-Id: I9b3ec6c101a89f30e61fc683808506c404e5474f
Partial-Bug: #1538506