Updated openstack/openstack Project: openstack/nova  b297e6b93d1a153d580b41256a655bc4e8925478
Aggregate object fixups

In the original removal of DictCompat from the Aggregate object, the
aggregate.get(foo, default) calls were changed to getattr(aggregate,
foo, default). This is an incorrect change, because the default of
getattr() is only used if AttributeError is raised. The original get()
call would return the default if the variable was not set, but in order
to get the default with getattr() 'in' has to be used. So now, the
getattr() calls are changed to:

foo = aggregate.foo if 'foo' in aggregate else default

Also, a comment was added for using getattr() on all aggregate fields,
explaining we can do the getattr() and not worry about metadata being
unset, because the compute API always sets it. nova.objects.base also
has a helper method to compare the primitives of two objects to check if
they're equal, so a helper method in the tests were changed over to use
that helper method.

Change-Id: Iee651704c90fcdda0938f907924a4565399601d7
Closes-Bug: #1554617