Updated openstack/openstack Project: openstack-infra/tripleo-ci  2a33763a7e838de3198ad6cb6abb11ce9b74976b
set -o pipefail in deploy.sh

If we don't do this, then commands that we pipe such as the image
build and undercloud install don't fail the run when they fail.
This causes unnecessary confusion and is a problem for testing
undercloud idempotency since a failure reinstalling the undercloud
may not cause a subsequent failure.

Change-Id: Ic4a39c60d3affc2e6234ffc0521c30288f1f95d5