Updated openstack/openstack Project: openstack/neutron  0e3c3e74ca686115d450a59a00ebfff01f4114b0
Ensures DNS_DRIVER is loaded before it is checked

Previously it was possible for DNS_DRIVER to be checked here before anything
had attempted to load it, causing the check to erroneuously fail. This patch
makes sure that the check will not fail simply because nothing had loaded it
prior by attempting to load it immediately before the check.

Change-Id: I34537beaf675db2634493dfef27b69051a8d0781
Closes-Bug: 1554151